API problem/AddRating

+++ frontend/database/schema.sql
@@ -167,7 +167,7 @@ CREATE TABLE IF NOT EXISTS `Contests` (
   `contestant_must_register`   tinyint(1) NOT NULL DEFAULT '0' COMMENT 'Indica que los participantes deben pre-registrarse antes de poder paticipar',
   `languages` set('c','cpp','java','py','rb','pl','cs','pas','kp','kj','cat','hs','cpp11') DEFAULT NULL COMMENT 'Un filtro (opcional) de qué lenguajes se pueden usar en un concurso',
   `recommended` BOOL NOT NULL DEFAULT  '0' COMMENT  'Mostrar el concurso en la lista de recomendados.',
-	`rating` float(2,2) NOT NULL DEFAULT '3.0' COMMENT 'Rating calulado para el concurso',
+	`rating` float NOT NULL DEFAULT '3.0' COMMENT 'Rating calulado para el concurso',
   PRIMARY KEY (`contest_id`),
   KEY `director_id` (`director_id`),
   KEY `rerun_id` (`contest_id`),
@@ -435,7 +435,7 @@ CREATE TABLE IF NOT EXISTS `Problems` (
   `slow` tinyint(1) NOT NULL DEFAULT 0,
   `deprecated` tinyint(1) NOT NULL DEFAULT 0,
   `email_clarifications` tinyint(1) NOT NULL DEFAULT 0,
-	`rating` float(2,2) NOT NULL DEFAULT '3.0',
+	`rating` float NOT NULL DEFAULT '3.0',
   PRIMARY KEY (`problem_id`),
   KEY `author_id` (`author_id`),
   UNIQUE KEY `problems_alias` (`alias`)
@@ -785,7 +785,7 @@ CREATE TABLE IF NOT EXISTS `User_Feedback_Rating` (
 		`entity_id` int(11) NOT NULL,
 		`feature_id` int(11) NOT NULL,
 		`create_date` timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-		`rating` float(2,2) NOT NULL,
+		`rating` float NOT NULL,
 		`comments` text,
 		PRIMARY KEY (`user_id`, `entity_id`, `feature_id`),
 		FOREIGN KEY (`user_id`) REFERENCES `Users` (`user_id`) ON DELETE NO ACTION ON UPDATE NO ACTION,
@@ -798,7 +798,7 @@ CREATE TABLE IF NOT EXISTS `User_Feedback_Rating` (
 CREATE TABLE IF NOT EXISTS `Entity_Feedback_Rating` (
 		`feature_id` int(11) NOT NULL,
 		`entity_id` int(11) NOT NULL,
-		`rating` float(2,2) NOT NULL,
+		`rating` float NOT NULL,
 		PRIMARY KEY (`feature_id`, `entity_id`),
 		FOREIGN KEY (`feature_id`) REFERENCES `Entity_Feedback_Features` (`feature_id`) ON DELETE NO ACTION ON UPDATE NO ACTION
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COMMENT='Guarda el rating calculado a cada entidad';
